feat: add auth flow and protected dashboard
