+++ src/proxy.ts
@@ -1,0 +1,11 @@
+import { withAuth } from "next-auth/middleware";
+
+export default withAuth({
+  pages: {
+    signIn: "/auth/login",
+  },
+});
+
+export const config = {
+  matcher: ["/dashboard/:path*"],
+};
