+++ src/shared/api/auth/index.ts
@@ -2,15 +2,25 @@ type BackendPostOptions = {
   headers?: Record<string, string>;
 };
 
+function joinUrl(baseUrl: string, path: string): string {
+  const normalizedBase = baseUrl.endsWith("/") ? baseUrl.slice(0, -1) : baseUrl;
+  const normalizedPath = path.startsWith("/") ? path : `/${path}`;
+
+  return `${normalizedBase}${normalizedPath}`;
+}
+
 export async function backendPost<TResponse>(
   path: string,
   body: unknown,
   options: BackendPostOptions = {},
 ): Promise<TResponse> {
   const baseUrl = process.env.BACKEND_URL;
-  if (!baseUrl) throw new Error("BACKEND_URL is not set");
 
-  const res = await fetch(`${baseUrl}${path}`, {
+  if (!baseUrl) {
+    throw new Error("BACKEND_URL is not set");
+  }
+
+  const res = await fetch(joinUrl(baseUrl, path), {
     method: "POST",
     headers: {
       "content-type": "application/json",
